+++ lib/xmat-time-input/ts/xmat-time-input.component.ts
@@ -1,10 +1,16 @@
 import {coerceBooleanProperty} from "@angular/cdk/coercion";
 import {FocusMonitor} from "@angular/cdk/a11y";
 import {
-    FormBuilder, FormControl, FormGroup, NG_VALIDATORS, NG_VALUE_ACCESSOR, ValidationErrors,
+    FormBuilder,
+    FormControl,
+    FormGroup,
+    NgControl,
+    NG_VALIDATORS,
+    NG_VALUE_ACCESSOR,
+    ValidationErrors,
     Validators
 } from "@angular/forms";
-import {Component, ElementRef, forwardRef, Input, OnDestroy, AfterViewInit, ViewEncapsulation} from "@angular/core";
+import {Component, ElementRef, forwardRef, Input, OnDestroy, ViewEncapsulation} from "@angular/core";
 import {MatFormFieldControl} from "@angular/material";
 import {Subject} from "rxjs/Subject";
 import {XmatTime} from "../../xmat-models/index";
@@ -16,7 +22,6 @@ const elementType = "input";
 const hoursPattern = new RegExp("^([01][0-9]|2[0-3])");
 const minutesPattern = new RegExp("^([0-5][0-9])$");
 
-//noinspection JSAnnotator
 @Component({
     selector: controlType,
     templateUrl: "../tpl/xmat-time-input.component.html",
@@ -27,8 +32,9 @@ const minutesPattern = new RegExp("^([0-5][0-9])$");
         {provide: NG_VALIDATORS, useExisting: forwardRef(() => XmatMatTimeInputComponent), multi: true}
     ],
     host: {
-        "[class.xmat-mat-time-input]": "true",
+        "[class.xmat-mat-time]": "true",
         "[class.floating]": "shouldLabelFloat",
+        "[class.empty-placeholder]": "!placeholder?.length",
         "[class.filled]": "isFilled",
         "[id]": "id",
         "[attr.aria-describedby]": "describedBy",
@@ -36,9 +42,9 @@ const minutesPattern = new RegExp("^([0-5][0-9])$");
     encapsulation: ViewEncapsulation.None
 })
 
-export class XmatMatTimeInputComponent implements MatFormFieldControl<XmatTime>, AfterViewInit, OnDestroy {
+export class XmatMatTimeInputComponent implements MatFormFieldControl<XmatTime>, OnDestroy {
 
-    static nextId = 0;
+    static nextId: number = 0;
 
     get empty() {
         const value = this.parts.value;
@@ -117,7 +123,7 @@ export class XmatMatTimeInputComponent implements MatFormFieldControl<XmatTime>,
     inputPlaceholder: string;
     invalid: boolean = false;
     model: XmatTime = new XmatTime();
-    ngControl = null;
+    ngControl: NgControl = null;
     parts: FormGroup;
     placeholderM: string = "";
     placeholderH: string = "";
@@ -155,15 +161,6 @@ export class XmatMatTimeInputComponent implements MatFormFieldControl<XmatTime>,
     ngOnDestroy(): void {
         this.stateChanges.complete();
         this._focusMonitor.stopMonitoring(this._elRef.nativeElement);
-    }
-
-    ngAfterViewInit(): void {
-        const xmatInput = this._elRef.nativeElement;
-        const container = xmatInput.closest("mat-form-field");
-        const isContainerValid = !!container && container.classList.contains("mat-form-field-type-xmat-time-input");
-        if (isContainerValid && container.contains(xmatInput)) {
-            this._$matFormField = container;
-        }
     }
 
     onContainerClick(event: MouseEvent): void {
@@ -245,16 +242,13 @@ export class XmatMatTimeInputComponent implements MatFormFieldControl<XmatTime>,
 
     }
 
-    validate(c: FormControl | FormGroup): ValidationErrors | null {
-        if (!!this._$matFormField) {
-            this.invalid = this.parts.invalid && this.parts.touched;
-            this._$matFormField.classList.toggle("mat-form-field-invalid", this.invalid);
-        }
+    validate(_c_: FormControl | FormGroup): ValidationErrors | null {
         const value = this.parts.value;
         const charsCount = value.hours.length + value.minutes.length;
         const addFormatError = this.parts.invalid && (charsCount > 0 || charsCount < 4);
-
+        this.errorState = addFormatError;
         return addFormatError ? {timeFormatError: true} : null;
+
     }
 
     // this is the initial value and reset are updated to the component