+++ src/app/services/public.ts
@@ -1,0 +1,5 @@
+export * from "./xmat-routes.service";
+export * from "./xmat-test-consts.service";
+export * from "./xmat-test-mock.service";
+export * from "./xmat-test-mocks-list.service";
+export * from "./xmat-test-rest.service";